% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{cce.loss}
\alias{cce.loss}
\title{Categorical Cross-Entropy Loss function, used for classification networks}
\usage{
cce.loss(A, Y)
}
\arguments{
\item{A}{a matrix of probabilities.}

\item{Y}{the true Y values.}
}
\value{
The floating-point loss for the network.
}
\description{
Categorical Cross-Entropy Loss function, used for classification networks
}
