% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.mlp.R
\name{create.mlp}
\alias{create.mlp}
\title{Initialize a Neural Network.}
\usage{
create.mlp(
  formula,
  data = NULL,
  hidden = c(),
  type = NULL,
  activation = "leaky.relu",
  valid.split = 0.25,
  randomize = TRUE,
  learning.rate = 0.001,
  momentum.beta = 0.9,
  rmsprop.beta = 0.999,
  batch.size = 64
)
}
\arguments{
\item{formula}{An object of class `formula` describing the model to be fit.}

\item{data}{A `data.frame` object containing the training and validation data.}

\item{hidden}{A vector indicating how many layers and hidden notes the network will have. Passing c(4,5,2) creates a network with 3 hidden layers with 4, 5, and 2 nodes, respectively.}

\item{type}{One of "regression", "logistic", or "classification".}

\item{activation}{The name of a function to be used for activations in the hidden layers. "leaky.relu", "relu", "tanh", and "sigmoid" are included in the package.}

\item{valid.split}{The fraction of the data to be used for validation.}

\item{randomize}{If TRUE, the data frame will be randomly sampled before train/validation split.
If not, the final rows are the validation data.}

\item{learning.rate}{The learning rate of the network.}

\item{momentum.beta}{Beta parameter for gradient descent with momentum.}

\item{batch.size}{Minibatch size. Choosing 1 is equivalent to stochastic gradient descent.}

\item{rmpsprop.beta}{Beta parameter for RMSprop gradient descent.}
}
\value{
An object of class `mlp`
}
\description{
Create a new neural network to fitting tabular data.

The network is initialized randomized weights and should be trained with the `train()` function.
* Input and output data are normalized, with factor variables one-hot encoded.
* Handles both regression and classification networks.
* Splits input dataframe into training and validation data.
* ADAM optimizer with minibatch normalization with tunable hyperparameters.
* Supports user-defined activation functions. See `relu` or `leaky.relu` for examples.
}
\examples{
 create.mlp(Species ~ ., data=iris, hidden=c(5,5,5), type="classification")
}
