% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.prop.R
\name{forward.prop}
\alias{forward.prop}
\title{Performs forward propagation through the neural network
Most users should use predict.mlp() as a wrapper}
\usage{
forward.prop(network, X = NULL, return.cache = FALSE)
}
\arguments{
\item{network}{An object of class `mlp` containing the neural network.}

\item{X}{The design matrix (if set to NULL), or a pre-processed matrix of input data.}

\item{return.cache}{if TRUE, instead of the final activation, the network returns all intermediate activations.
Used for backwards propagation.}
}
\value{
The activations of the final layer of the network. If return.cache is TRUE, then all intermediate layer activations are instead returned in a list.
}
\description{
Performs forward propagation through the neural network
Most users should use predict.mlp() as a wrapper
}
