% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{leaky.relu}
\alias{leaky.relu}
\title{A "leaky" RELU function, and derivative. The slope in the negative region is 0.1.}
\usage{
leaky.relu(X, deriv = FALSE)
}
\arguments{
\item{X}{A matrix.}

\item{deriv}{Set this to TRUE if you want the derivative instead of the function.}
}
\value{
A matrix representing the leaky RELU(X), or its derivative.
}
\description{
A "leaky" RELU function, and derivative. The slope in the negative region is 0.1.
}
