% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{linear}
\alias{linear}
\title{A linear activation function, and its derivative.}
\usage{
linear(X, deriv = FALSE)
}
\arguments{
\item{X}{A matrix.}

\item{deriv}{Set this to TRUE if you want the derivative instead of the function.}
}
\value{
X, or its derivative (a matrix of all 1s).
}
\description{
A linear activation function, and its derivative.
}
