% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{logit.loss}
\alias{logit.loss}
\title{A Logistic Loss function, used for binary classification networks.}
\usage{
logit.loss(A, Y)
}
\arguments{
\item{A}{a vector of predicted Y values.}

\item{Y}{the true Y values.}
}
\value{
The floating-point loss for the network.
}
\description{
A Logistic Loss function, used for binary classification networks.
}
