% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{mse.loss}
\alias{mse.loss}
\title{A Mean Squared Error Loss function, used for regression networks.}
\usage{
mse.loss(A, Y)
}
\arguments{
\item{A}{a vector of predicted Y values.}

\item{Y}{the true Y values.}
}
\value{
The floating-point loss for the network.
}
\description{
A Mean Squared Error Loss function, used for regression networks.
}
