% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mlp.R
\name{plot.mlp}
\alias{plot.mlp}
\title{Plot the Loss or Other Metrics of a Neural Network}
\usage{
\method{plot}{mlp}(network, metric = "loss", ylim = NULL)
}
\arguments{
\item{network}{An object of class `mlp` to be plotted.}

\item{metric}{One of "loss", "accuracy", or "residuals" to be plotted vs. training epoch.}

\item{ylim}{Y-Axis limits for the plot.}
}
\description{
Plot the Loss or Other Metrics of a Neural Network
}
