% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mlp.R
\name{predict.mlp}
\alias{predict.mlp}
\title{Make predictions using the neural network.}
\usage{
\method{predict}{mlp}(network, newdata = NULL, type = "numeric")
}
\arguments{
\item{network}{An object of class `mlp` used to make predictions.}

\item{newdata}{An object of class `data.frame` with the same input columns as the data frame used to create the network.
If left NULL, the training data will be used.
If set to "valid", the validation data will be used.}

\item{type}{If set to "numeric", then predicted numeric values (regression), probabilities (logistic), or a matrix of probability for each class (classification) is returned. For a classification network, type="labels" can be used to return the class with the highest probability.}
}
\value{
A vector (or matrix, for a classification network) of the predicted outputs.
}
\description{
Make predictions using the neural network.
}
\examples{
   predict(network, type="labels")
   predict(network, newdata="valid", type="labels")
   predict(network, newdata=test.data)
}
