% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{relu}
\alias{relu}
\title{A rectified linear unit or RELU = max(0, x) function, and its derivative.}
\usage{
relu(X, deriv = FALSE)
}
\arguments{
\item{X}{A matrix.}

\item{deriv}{Set this to TRUE if you want the derivative instead of the function.}
}
\value{
A matrix representing the RELU(X), or its derivative.
}
\description{
A rectified linear unit or RELU = max(0, x) function, and its derivative.
}
