% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{sigmoid}
\alias{sigmoid}
\title{A sigmoid (= logistic) function, and its derivative.}
\usage{
sigmoid(X, deriv = FALSE)
}
\arguments{
\item{X}{A matrix.}

\item{deriv}{Set this to TRUE if you want the derivative instead of the function.}
}
\value{
A matrix representing the sigmoid(X), or its derivative.
}
\description{
A sigmoid (= logistic) function, and its derivative.
}
