% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{softmax}
\alias{softmax}
\title{A softmax activation, used for classifier networks as the output layer.}
\usage{
softmax(X, deriv = FALSE)
}
\arguments{
\item{X}{A matrix.}

\item{deriv}{Set this to TRUE if you want the derivative instead of the function.}
}
\value{
An matrix representing the probability of each class for each observation.
}
\description{
A softmax activation, used for classifier networks as the output layer.
}
