% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.dataset.R
\name{split.dataset}
\alias{split.dataset}
\title{Split a dataset into a design (X) matrix and labels (Y), and one-hot encodes factors. Does not need to be called by the user.}
\usage{
\method{split}{dataset}(formula, data)
}
\arguments{
\item{formula}{An object of class `formula` describing the model to be fit, to determine X and Y.}

\item{data}{A `data.frame` containing the dataset to be split.}
}
\value{
A list containing the design matrix (X), and the Y values (one-hot encoded as necessary).
    If Y is a factor, then raw.Y is still factors, and levels contains the levels that the factor can take.
}
\description{
Split a dataset into a design (X) matrix and labels (Y), and one-hot encodes factors. Does not need to be called by the user.
}
