% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations-losses.R
\name{tanh}
\alias{tanh}
\title{A hyperbolic tangent function, and its derivative.}
\usage{
tanh(X, deriv = FALSE)
}
\arguments{
\item{X}{A matrix.}

\item{deriv}{Set this to TRUE if you want the derivative instead of the function.}
}
\value{
A matrix representing the tanh(X), or its derivative.
}
\description{
A hyperbolic tangent function, and its derivative.
}
