% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train a Neural Network}
\usage{
train(network, epochs = 1000)
}
\arguments{
\item{network}{An object of class `mlp` to be trained.}

\item{epochs}{The number of epochs to train for.}
}
\value{
An object of class `mlp` representing the input network after training.
}
\description{
Train a Neural Network
}
\examples{
    network <- train(network, 100)
    net.trained <- train(net, 100)
}
